# 记忆模块/proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/memoryRecords:
        get:
            tags:
                - MemoryService
            operationId: MemoryService_ListMemoryRecords
            parameters:
                - name: namespace
                  in: query
                  schema:
                    type: string
                - name: scopes
                  in: query
                  schema:
                    type: array
                    items:
                        enum:
                            - MEMORY_SCOPE_UNSPECIFIED
                            - MEMORY_SCOPE_USER
                            - MEMORY_SCOPE_PROJECT
                            - MEMORY_SCOPE_SESSION
                            - MEMORY_SCOPE_ORGANIZATION
                        type: string
                        format: enum
                - name: kinds
                  in: query
                  schema:
                    type: array
                    items:
                        enum:
                            - MEMORY_KIND_UNSPECIFIED
                            - MEMORY_KIND_PROFILE
                            - MEMORY_KIND_PREFERENCE
                            - MEMORY_KIND_BEHAVIOR
                            - MEMORY_KIND_PROJECT_FACT
                            - MEMORY_KIND_ANALYSIS
                            - MEMORY_KIND_DECISION
                            - MEMORY_KIND_CORRECTION
                            - MEMORY_KIND_FOLLOW_UP
                            - MEMORY_KIND_RUN_SUMMARY
                        type: string
                        format: enum
                - name: statuses
                  in: query
                  schema:
                    type: array
                    items:
                        enum:
                            - MEMORY_STATUS_UNSPECIFIED
                            - MEMORY_STATUS_ACTIVE
                            - MEMORY_STATUS_PENDING
                            - MEMORY_STATUS_SUPERSEDED
                            - MEMORY_STATUS_ARCHIVED
                        type: string
                        format: enum
                - name: scopeId
                  in: query
                  schema:
                    type: string
                - name: query
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListMemoryRecordsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /api/v1/memoryRecords:purge:
        post:
            tags:
                - MemoryService
            operationId: MemoryService_PurgeMemoryRecords
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PurgeMemoryRecordsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PurgeMemoryRecordsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/memoryRecords:upsert:
        post:
            tags:
                - MemoryService
            operationId: MemoryService_UpsertMemoryRecord
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpsertMemoryRecordRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MemoryRecord'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ListMemoryRecordsResponse:
            type: object
            properties:
                memoryRecords:
                    type: array
                    items:
                        $ref: '#/components/schemas/MemoryRecord'
                nextPageToken:
                    type: string
...
        MemoryEvidence:
            required:
                - sourceType
                - sourceRef
                - quote
                - observedTime
            type: object
            properties:
                sourceType:
                    type: string
                sourceRef:
                    type: string
                quote:
                    type: string
                observedTime:
                    type: string
                    format: date-time
                weight:
                    type: number
                    format: double
                fingerprint:
                    readOnly: true
                    type: string
        MemoryRecord:
            required:
                - namespace
                - scope
                - kind
                - key
                - value
                - origin
                - status
            type: object
            properties:
                name:
                    type: string
                namespace:
                    type: string
                scope:
                    enum:
                        - MEMORY_SCOPE_UNSPECIFIED
                        - MEMORY_SCOPE_USER
                        - MEMORY_SCOPE_PROJECT
                        - MEMORY_SCOPE_SESSION
                        - MEMORY_SCOPE_ORGANIZATION
                    type: string
                    format: enum
                scopeId:
                    type: string
                kind:
                    enum:
                        - MEMORY_KIND_UNSPECIFIED
                        - MEMORY_KIND_PROFILE
                        - MEMORY_KIND_PREFERENCE
                        - MEMORY_KIND_BEHAVIOR
                        - MEMORY_KIND_PROJECT_FACT
                        - MEMORY_KIND_ANALYSIS
                        - MEMORY_KIND_DECISION
                        - MEMORY_KIND_CORRECTION
                        - MEMORY_KIND_FOLLOW_UP
                        - MEMORY_KIND_RUN_SUMMARY
                    type: string
                    format: enum
                key:
                    type: string
                value:
                    type: string
                summary:
                    type: string
                origin:
                    enum:
                        - MEMORY_ORIGIN_UNSPECIFIED
                        - MEMORY_ORIGIN_EXPLICIT
                        - MEMORY_ORIGIN_INFERRED
                        - MEMORY_ORIGIN_SYSTEM
                        - MEMORY_ORIGIN_MANUAL
                    type: string
                    format: enum
                status:
                    enum:
                        - MEMORY_STATUS_UNSPECIFIED
                        - MEMORY_STATUS_ACTIVE
                        - MEMORY_STATUS_PENDING
                        - MEMORY_STATUS_SUPERSEDED
                        - MEMORY_STATUS_ARCHIVED
                    type: string
                    format: enum
                confidence:
                    type: number
                    format: double
                importance:
                    type: number
                    format: double
                sensitive:
                    type: boolean
                evidenceCount:
                    readOnly: true
                    type: integer
                    format: int32
                version:
                    readOnly: true
                    type: integer
                    format: int32
                createTime:
                    readOnly: true
                    type: string
                    format: date-time
                updateTime:
                    readOnly: true
                    type: string
                    format: date-time
                lastConfirmedTime:
                    type: string
                    format: date-time
                expireTime:
                    type: string
                    format: date-time
                evidence:
                    readOnly: true
                    type: array
                    items:
                        $ref: '#/components/schemas/MemoryEvidence'
                revisions:
                    readOnly: true
                    type: array
                    items:
                        $ref: '#/components/schemas/MemoryRevision'
...
        PurgeMemoryRecordsRequest:
            required:
                - namespace
            type: object
            properties:
                namespace:
                    type: string
        PurgeMemoryRecordsResponse:
            type: object
            properties:
                deletedCount:
                    type: integer
                    format: int32
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        UpsertMemoryRecordRequest:
            required:
                - memoryRecord
            type: object
            properties:
                memoryRecord:
                    $ref: '#/components/schemas/MemoryRecord'
                evidence:
                    $ref: '#/components/schemas/MemoryEvidence'
                expectedVersion:
                    type: integer
                    format: int32
                reason:
                    type: string
...